refactor(adaptive-ui): align app bar ownership
