+++ lib/pages/habit_detail/widgets.dart
@@ -16,6 +16,7 @@ import '../../widgets/widgets.dart' show HabitDivider;
 
 export '_widgets/habit_desc_cell_tile.dart';
 export '_widgets/habit_detail_appbar.dart';
+export '_widgets/habit_detail_appbar_actions.dart';
 export '_widgets/habit_detail_chart_title.dart';
 export '_widgets/habit_detail_desc_tile.dart';
 export '_widgets/habit_detail_fab.dart';
