+++ packages/mhabit_adaptive_ui/lib/src/adaptive/adaptive_app_bar_actions.dart
@@ -1,9 +1,10 @@
 import 'package:adaptive_actions/cupertino.dart';
 import 'package:adaptive_actions/material.dart';
-import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 
 import '../adaptive_style.dart';
+import '../cupertino/cupertino_app_bar_actions.dart';
+import '../material/material_app_bar_actions.dart';
 
 /// Receives an invoked action together with the button that anchored it.
 ///
@@ -53,13 +54,13 @@ class AdaptiveAppBarActions<T extends Object> extends StatelessWidget {
     required this.primaryCapacity,
     this.maxPrimaryActions,
     this.materialIconBuilder,
-    this.appleIconBuilder,
     this.materialOverflowIcon,
-    this.appleOverflowIcon,
     this.overflowTooltip,
     this.primaryActionDecorator,
   }) : assert(primaryCapacity >= 0 && primaryCapacity < double.infinity),
        assert(maxPrimaryActions == null || maxPrimaryActions >= 0),
+       appleIconBuilder = null,
+       appleOverflowIcon = null,
        style = AdaptiveStyle.material;
 
   const AdaptiveAppBarActions.apple({
@@ -68,14 +69,14 @@ class AdaptiveAppBarActions<T extends Object> extends StatelessWidget {
     required this.onInvoke,
     required this.primaryCapacity,
     this.maxPrimaryActions,
-    this.materialIconBuilder,
     this.appleIconBuilder,
-    this.materialOverflowIcon,
     this.appleOverflowIcon,
     this.overflowTooltip,
     this.primaryActionDecorator,
   }) : assert(primaryCapacity >= 0 && primaryCapacity < double.infinity),
        assert(maxPrimaryActions == null || maxPrimaryActions >= 0),
+       materialIconBuilder = null,
+       materialOverflowIcon = null,
        style = AdaptiveStyle.apple;
 
   final AdaptiveStyle? style;
@@ -102,87 +103,26 @@ class AdaptiveAppBarActions<T extends Object> extends StatelessWidget {
         materialLocalizations?.showMenuTooltip ??
         'More actions';
     return switch (effectiveStyle) {
-      AdaptiveStyle.material => _buildMaterial(
-        context,
-        effectiveOverflowTooltip,
+      AdaptiveStyle.material => MaterialAppBarActions<T>(
+        collection: collection,
+        onInvoke: onInvoke,
+        primaryCapacity: primaryCapacity,
+        maxPrimaryActions: maxPrimaryActions,
+        iconBuilder: materialIconBuilder,
+        overflowIcon: materialOverflowIcon,
+        overflowTooltip: effectiveOverflowTooltip,
+        primaryActionDecorator: primaryActionDecorator,
       ),
-      AdaptiveStyle.apple => _buildApple(context, effectiveOverflowTooltip),
+      AdaptiveStyle.apple => CupertinoAppBarActions<T>(
+        collection: collection,
+        onInvoke: onInvoke,
+        primaryCapacity: primaryCapacity,
+        maxPrimaryActions: maxPrimaryActions,
+        iconBuilder: appleIconBuilder,
+        overflowIcon: appleOverflowIcon,
+        overflowTooltip: effectiveOverflowTooltip,
+        primaryActionDecorator: primaryActionDecorator,
+      ),
     };
-  }
-
-  Widget _buildMaterial(BuildContext context, String overflowTooltip) {
-    final primaryAnchors = <T, BuildContext>{};
-    BuildContext? overflowAnchor;
-    return MaterialAdaptiveActions<T>.moreAction(
-      actions: collection,
-      onInvoke: (value) =>
-          onInvoke(primaryAnchors[value] ?? overflowAnchor ?? context, value),
-      primaryCapacity: primaryCapacity,
-      maxPrimaryActions: maxPrimaryActions,
-      iconBuilder: materialIconBuilder,
-      actionButtonBuilder: (context, action, onPressed, defaultBuilder) {
-        return Builder(
-          builder: (anchorContext) {
-            final payload = action.payload;
-            if (payload != null) primaryAnchors[payload] = anchorContext;
-            final child = defaultBuilder(anchorContext, action, onPressed);
-            return primaryActionDecorator?.call(anchorContext, action, child) ??
-                child;
-          },
-        );
-      },
-      overflowButtonBuilder: (context, onPressed, defaultBuilder) {
-        return Builder(
-          builder: (anchorContext) {
-            overflowAnchor = anchorContext;
-            return defaultBuilder(anchorContext, onPressed);
-          },
-        );
-      },
-      overflowIcon: materialOverflowIcon ?? const Icon(Icons.more_vert),
-      overflowTooltip: overflowTooltip,
-      presentationOverride: MaterialActionPresentation.iconOnly,
-      style: const MaterialAdaptiveActionsStyle(iconSize: 24),
-      fadeDuration: Duration.zero,
-      resizeDuration: Duration.zero,
-    );
-  }
-
-  Widget _buildApple(BuildContext context, String overflowTooltip) {
-    final primaryAnchors = <T, BuildContext>{};
-    BuildContext? overflowAnchor;
-    return CupertinoAdaptiveActions<T>.moreAction(
-      actions: collection,
-      onInvoke: (value) =>
-          onInvoke(primaryAnchors[value] ?? overflowAnchor ?? context, value),
-      primaryCapacity: primaryCapacity,
-      maxPrimaryActions: maxPrimaryActions,
-      iconBuilder: appleIconBuilder,
-      actionButtonBuilder: (context, action, onPressed, defaultBuilder) {
-        return Builder(
-          builder: (anchorContext) {
-            final payload = action.payload;
-            if (payload != null) primaryAnchors[payload] = anchorContext;
-            final child = defaultBuilder(anchorContext, action, onPressed);
-            return primaryActionDecorator?.call(anchorContext, action, child) ??
-                child;
-          },
-        );
-      },
-      overflowButtonBuilder: (context, onPressed, defaultBuilder) {
-        return Builder(
-          builder: (anchorContext) {
-            overflowAnchor = anchorContext;
-            return defaultBuilder(anchorContext, onPressed);
-          },
-        );
-      },
-      overflowIcon: appleOverflowIcon ?? const Icon(CupertinoIcons.ellipsis),
-      overflowTooltip: overflowTooltip,
-      presentationOverride: CupertinoActionPresentation.iconOnly,
-      invokeAfterMenuClosed: true,
-      fadeDuration: Duration.zero,
-      resizeDuration: Duration.zero,
-    );
   }
 }
