+++ packages/mhabit_adaptive_ui/lib/src/adaptive/adaptive_sliver_app_bar.dart
@@ -1,109 +1,17 @@
 import 'package:flutter/material.dart';
 
 import '../adaptive_style.dart';
+import '../cupertino/app_bar_apple_style.dart';
 import '../cupertino/cupertino_sliver_app_bar.dart';
-import '../window_control/material_app_bar.dart';
-import '../window_control/toolbar_geometry.dart';
-import 'app_bar_apple_style.dart';
+import '../material/app_bar_material_style.dart';
+import '../material/material_sliver_app_bar.dart';
 
-export 'app_bar_apple_style.dart' show AppBarAppleStyle;
+export '../cupertino/app_bar_apple_style.dart' show AppBarAppleStyle;
+export '../material/app_bar_material_style.dart' show AppBarMaterialStyle;
 
 const List<Widget> _kDefaultActions = <Widget>[];
 
-/// Style config for the Material branch of [AdaptiveSliverAppBar].
-///
-/// App-bar fields map to [SliverAppBar]. [windowControlEdgePadding] is the
-/// Material visual baseline added only on sides with window-control avoidance.
-class AppBarMaterialStyle {
-  const AppBarMaterialStyle({
-    this.centerTitle = true,
-    this.floating = true,
-    this.snap = true,
-    this.pinned = true,
-    this.forceElevated = false,
-    this.scrolledUnderElevation,
-    this.shadowColor = Colors.transparent,
-    this.bottom,
-    this.toolbarHeight,
-    this.windowControlEdgePadding = materialWindowControlEdgePadding,
-  });
-
-  final bool centerTitle;
-  final bool floating;
-  final bool snap;
-  final bool pinned;
-  final bool forceElevated;
-  final double? scrolledUnderElevation;
-  final Color? shadowColor;
-  final PreferredSizeWidget? bottom;
-
-  /// Material-only toolbar height, overriding the shared app-bar height.
-  final double? toolbarHeight;
-
-  /// {@macro mhabit.windowControlEdgePadding}
-  final EdgeInsetsDirectional windowControlEdgePadding;
-
-  AppBarMaterialStyle copyWith({
-    bool? centerTitle,
-    bool? floating,
-    bool? snap,
-    bool? pinned,
-    bool? forceElevated,
-    double? scrolledUnderElevation,
-    Color? shadowColor,
-    PreferredSizeWidget? bottom,
-    double? toolbarHeight,
-    EdgeInsetsDirectional? windowControlEdgePadding,
-  }) => AppBarMaterialStyle(
-    centerTitle: centerTitle ?? this.centerTitle,
-    floating: floating ?? this.floating,
-    snap: snap ?? this.snap,
-    pinned: pinned ?? this.pinned,
-    forceElevated: forceElevated ?? this.forceElevated,
-    scrolledUnderElevation:
-        scrolledUnderElevation ?? this.scrolledUnderElevation,
-    shadowColor: shadowColor ?? this.shadowColor,
-    bottom: bottom ?? this.bottom,
-    toolbarHeight: toolbarHeight ?? this.toolbarHeight,
-    windowControlEdgePadding:
-        windowControlEdgePadding ?? this.windowControlEdgePadding,
-  );
-
-  @override
-  bool operator ==(Object other) =>
-      other is AppBarMaterialStyle &&
-      other.centerTitle == centerTitle &&
-      other.floating == floating &&
-      other.snap == snap &&
-      other.pinned == pinned &&
-      other.forceElevated == forceElevated &&
-      other.scrolledUnderElevation == scrolledUnderElevation &&
-      other.shadowColor == shadowColor &&
-      other.bottom == bottom &&
-      other.toolbarHeight == toolbarHeight &&
-      other.windowControlEdgePadding == windowControlEdgePadding;
-
-  @override
-  int get hashCode => Object.hash(
-    centerTitle,
-    floating,
-    snap,
-    pinned,
-    forceElevated,
-    scrolledUnderElevation,
-    shadowColor,
-    bottom,
-    toolbarHeight,
-    windowControlEdgePadding,
-  );
-}
-
 /// Per-style config overrides for [AdaptiveSliverAppBar].
-///
-/// Callers select styles by name (no platform branching): the dispatched
-/// style consumes its matching config, the other one is ignored. Null means
-/// "all defaults"; partial overrides fill the rest from the config
-/// constructor defaults.
 class AppBarStyles {
   const AppBarStyles({this.material, this.apple});
 
@@ -137,8 +45,8 @@ class AppBarStyles {
 /// (`CupertinoSliverNavigationBar`).
 ///
 /// Shared parameters live at the top level; style-divergent knobs live in
-/// [AppBarStyles]. Material resolves [height] as [SliverAppBar.toolbarHeight].
-/// Apple uses a fixed, centered toolbar when [height] is provided unless
+/// [styles]. Material resolves [height] as [SliverAppBar.toolbarHeight]. Apple
+/// uses a fixed, centered toolbar when [height] is provided unless
 /// [AppBarAppleStyle.collapsible] requests the native collapsing navigation
 /// bar behavior. A shared [bottom] is hosted below the fixed toolbar on both
 /// renderers; a Material-only [AppBarMaterialStyle.bottom] overrides it.
@@ -153,7 +61,7 @@ class AdaptiveSliverAppBar extends StatelessWidget {
     this.bottom,
     this.styles,
   }) : assert(bottom == null || height != null),
-       style = null;
+       _adaptiveStyle = null;
 
   const AdaptiveSliverAppBar.material({
     super.key,
@@ -164,7 +72,7 @@ class AdaptiveSliverAppBar extends StatelessWidget {
     this.height,
     this.bottom,
     this.styles,
-  }) : style = AdaptiveStyle.material;
+  }) : _adaptiveStyle = AdaptiveStyle.material;
 
   const AdaptiveSliverAppBar.apple({
     super.key,
@@ -176,9 +84,9 @@ class AdaptiveSliverAppBar extends StatelessWidget {
     this.bottom,
     this.styles,
   }) : assert(bottom == null || height != null),
-       style = AdaptiveStyle.apple;
+       _adaptiveStyle = AdaptiveStyle.apple;
 
-  final AdaptiveStyle? style;
+  final AdaptiveStyle? _adaptiveStyle;
   final Widget title;
   final List<Widget> actions;
   final Widget? leading;
@@ -195,38 +103,19 @@ class AdaptiveSliverAppBar extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    final effectiveStyle = style ?? AdaptiveStyle.of(context);
+    final effectiveStyle = _adaptiveStyle ?? AdaptiveStyle.of(context);
     return switch (effectiveStyle) {
-      AdaptiveStyle.material => _buildMaterial(_effectiveMaterialStyle),
+      AdaptiveStyle.material => MaterialSliverAppBar(
+        title: title,
+        actions: actions,
+        leading: leading,
+        onLeadingPressed: onLeadingPressed,
+        height: height,
+        bottom: bottom,
+        style: _effectiveMaterialStyle,
+      ),
       AdaptiveStyle.apple => _buildApple(_effectiveAppleStyle),
     };
-  }
-
-  Widget _buildMaterial(AppBarMaterialStyle config) {
-    // Equivalent of the app's current `SliverTopAppBar` baseline.
-    return WindowControlSliverAppBar(
-      floating: config.floating,
-      snap: config.snap,
-      pinned: config.pinned,
-      centerTitle: config.centerTitle,
-      toolbarHeight: config.toolbarHeight ?? height ?? kToolbarHeight,
-      forceElevated: config.forceElevated,
-      scrolledUnderElevation: config.scrolledUnderElevation,
-      shadowColor: config.shadowColor,
-      bottom: config.bottom ?? bottom,
-      title: title,
-      automaticallyImplyLeading: leading == null && onLeadingPressed == null,
-      leading:
-          leading ??
-          (onLeadingPressed == null
-              ? null
-              : IconButton(
-                  onPressed: onLeadingPressed,
-                  icon: const Icon(Icons.arrow_back),
-                )),
-      actions: actions.isEmpty ? null : actions,
-      windowControlEdgePadding: config.windowControlEdgePadding,
-    );
   }
 
   Widget _buildApple(AppBarAppleStyle effectiveStyle) => CupertinoSliverAppBar(
