+++ packages/mhabit_adaptive_ui/lib/src/cupertino/cupertino_sliver_search_bar.dart
@@ -4,7 +4,6 @@ import 'package:adaptive_actions/cupertino.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart' show Easing;
 
-import '../adaptive/app_bar_apple_style.dart';
 import '../breakpoints/breakpoints.dart';
 import '../breakpoints/window_size_class.dart';
 import '../shell/navigation_sidebar_app_bar_leading.dart';
@@ -111,7 +110,6 @@ class CupertinoSliverSearchBar extends StatefulWidget {
     this.bottom,
     this.bottomExtent = 0.0,
     this.pinned = true,
-    this.style,
   }) : assert(bottomExtent >= 0.0),
        assert(bottom != null || bottomExtent == 0.0);
 
@@ -136,9 +134,6 @@ class CupertinoSliverSearchBar extends StatefulWidget {
   final Widget? bottom;
   final double bottomExtent;
   final bool pinned;
-  final AppBarAppleStyle? style;
-
-  AppBarAppleStyle get _effectiveStyle => style ?? const AppBarAppleStyle();
 
   @override
   State<CupertinoSliverSearchBar> createState() =>
@@ -244,7 +239,6 @@ class _CupertinoSliverSearchBarState extends State<CupertinoSliverSearchBar> {
 
   @override
   Widget build(BuildContext context) {
-    final effectiveStyle = widget._effectiveStyle;
     final sidebarLeading = NavigationSidebarAppBarLeading.maybeOf(context);
     final screenWidth = MediaQuery.sizeOf(context).width;
     final widthClass = Breakpoints.of(context).widthClass(screenWidth);
@@ -268,12 +262,11 @@ class _CupertinoSliverSearchBarState extends State<CupertinoSliverSearchBar> {
           child: Stack(
             fit: StackFit.expand,
             children: [
-              WindowControlCupertinoNavigationBar(
+              const WindowControlCupertinoNavigationBar(
                 automaticallyImplyLeading: false,
                 transitionBetweenRoutes: false,
-                automaticBackgroundVisibility:
-                    effectiveStyle.automaticBackgroundVisibility,
-                backgroundColor: effectiveStyle.backgroundColor,
+                automaticBackgroundVisibility: true,
+                backgroundColor: CupertinoColors.transparent,
                 border: null,
               ),
               Positioned(
