+++ packages/mhabit_adaptive_ui/lib/src/material/material_app_bar_actions.dart
@@ -1,0 +1,70 @@
+import 'package:adaptive_actions/material.dart';
+import 'package:flutter/material.dart';
+
+/// Material renderer adapter for adaptive app-bar actions.
+class MaterialAppBarActions<T extends Object> extends StatelessWidget {
+  const MaterialAppBarActions({
+    super.key,
+    required this.collection,
+    required this.onInvoke,
+    required this.primaryCapacity,
+    required this.overflowTooltip,
+    this.maxPrimaryActions,
+    this.iconBuilder,
+    this.overflowIcon,
+    this.primaryActionDecorator,
+  });
+
+  final ActionCollection<T> collection;
+  final void Function(BuildContext anchorContext, T value) onInvoke;
+  final double primaryCapacity;
+  final int? maxPrimaryActions;
+  final MaterialActionIconBuilder<T>? iconBuilder;
+  final Widget? overflowIcon;
+  final String overflowTooltip;
+  final Widget Function(
+    BuildContext context,
+    AdaptiveAction<T> action,
+    Widget child,
+  )?
+  primaryActionDecorator;
+
+  @override
+  Widget build(BuildContext context) {
+    final primaryAnchors = <T, BuildContext>{};
+    BuildContext? overflowAnchor;
+    return MaterialAdaptiveActions<T>.moreAction(
+      actions: collection,
+      onInvoke: (value) =>
+          onInvoke(primaryAnchors[value] ?? overflowAnchor ?? context, value),
+      primaryCapacity: primaryCapacity,
+      maxPrimaryActions: maxPrimaryActions,
+      iconBuilder: iconBuilder,
+      actionButtonBuilder: (context, action, onPressed, defaultBuilder) {
+        return Builder(
+          builder: (anchorContext) {
+            final payload = action.payload;
+            if (payload != null) primaryAnchors[payload] = anchorContext;
+            final child = defaultBuilder(anchorContext, action, onPressed);
+            return primaryActionDecorator?.call(anchorContext, action, child) ??
+                child;
+          },
+        );
+      },
+      overflowButtonBuilder: (context, onPressed, defaultBuilder) {
+        return Builder(
+          builder: (anchorContext) {
+            overflowAnchor = anchorContext;
+            return defaultBuilder(anchorContext, onPressed);
+          },
+        );
+      },
+      overflowIcon: overflowIcon ?? const Icon(Icons.more_vert),
+      overflowTooltip: overflowTooltip,
+      presentationOverride: MaterialActionPresentation.iconOnly,
+      style: const MaterialAdaptiveActionsStyle(iconSize: 24),
+      fadeDuration: Duration.zero,
+      resizeDuration: Duration.zero,
+    );
+  }
+}
